fix: move package to internal
