+++ internal/schedule/interval.go
